Add files via upload
Python application for the classic monty_hall game, using SQL in the backend. This applications creates a friendly user-interface to play the game, view the statistics of a player and coming to the technical thing, we can find out the "conditional probability" which is the critical logic that is underlying in this game which puzzled many mathematicians in the past.